+++ packages/kobber-components/src/radio/radio-group/RadioGroup.ts
@@ -13,7 +13,7 @@ import {
 } from "../../base/internal/form";
 import componentStyles from "../../base/styles/component.styles";
 import type { RadioInput } from "../radio-input/RadioInput";
-import { radioGroupName, radioInputName, GroupProps } from "../Radio.core";
+import { radioGroupName, radioInputName, type GroupProps } from "../Radio.core";
 import { radioGroupStyles } from "./RadioGroup.styles";
 import { customElement } from "../../base/utilities/customElementDecorator";
 
@@ -111,7 +111,7 @@ export class RadioGroup extends ShoelaceElement implements Props {
     super.connectedCallback();
     this.value = this.currentValue;
     // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-ignore
+    // @ts-expect-error
     navigation.addEventListener("navigate", () => {
       // Experimental functionality, does not work in Firefox. Might change in the future.
       this.handleUrlChange();