+++ packages/kobber-icons/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @gyldendal/kobber-icons
 
+## 0.6.6
+
+### Patch Changes
+
+- New components Display, Title, TextLabel and TextBody
+- Updated dependencies
+  - @gyldendal/kobber-base@0.6.6
+
 ## 0.6.5
 
 ### Patch Changes